requirements-dev.txt
# Development dependencies
-r requirements.txt

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0

# Code quality
black>=23.0.0
isort>=5.12.0
mypy>=1.4.0
flake8>=6.0.0
pre-commit>=3.3.0

# Type stubs
types-python-dateutil>=2.8.0
types-pytz>=2023.3.0
sqlalchemy-stubs>=0.4

# Documentation
sphinx>=7.0.0
sphinx-rtd-theme>=1.3.0
sphinx-autodoc-typehints>=1.24.0
